Finished searching connections for osmawalk.

+++ raptor/structs.h
@@ -125,13 +125,13 @@ struct mem_data {
 //*********************************/
 
 struct stop_arr{
-	uint32_t to;
+	uint64_t to;
 	uint64_t arrival;
 };
 
 struct stop_route{
 	uint64_t departure;
-	uint32_t id;
+	uint64_t id;
 	uint32_t n_stops;
 	struct stop_arr * stops;
 };
